refactor: change structure project
